requirements.txt
# Production Dependencies for API Gateway
# These are the only dependencies required for running the API Gateway in production.
# Install with: pip install -r requirements.txt

# Web Framework & Server
fastapi==0.110.0        # Async web framework with automatic OpenAPI generation
uvicorn[standard]==0.29.0  # ASGI server with performance optimizations

# HTTP Client
httpx==0.27.0           # Async HTTP client for downstream service communication

# Data Validation & Settings
pydantic==2.6.4         # Data validation using Python type hints
pydantic-settings==2.2.1  # Settings management from environment variables

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT token validation and signing

# Distributed Systems & Resilience
redis==5.0.1            # Redis client for distributed rate limiting (optional)
tenacity==8.2.3         # Retry logic with exponential backoff for circuit breaker

# Monitoring & Observability
prometheus-client==0.19.0  # Prometheus metrics exporter

# Rate Limiting
slowapi==0.1.9          # Rate limiting library (fallback to custom implementation)

# Configuration Management
PyYAML==6.0.1           # YAML file parsing for configuration
